==> 2 Analysis/Phosphoproteome/Combined/6p Volcano plotter.Rmd <==
---
title: "Volcano plotting"
output: 
---

```{r}
#######################################################################################################################################
# Combined analysis Notebook 6 - DESCRIPTION
#
# Here we make a volcano plot to visualise overall changes in phosphorylation.
# In particular, we highlight proteins of interest by configuring the EnhancedVolcano function.
# Note that it is able to take some standard ggplot commands.
# Some of our configurations are not documented by the official documentation - these were discovered by blood, sweat and tears!
# 
#
# INPUTS = results of mutliple t tests with BH correction (.csv)
# OUTPUTS = volcano plots (.pdf)
# 
#######################################################################################################################################
```

```{r}
# Setup
# Clear your workspace
rm(list=ls())

# Load useful packages
if (!require("tidyverse")) {
  install.packages("tidyverse", dependencies = TRUE)
  library(tidyverse)
}

if (!require("BiocManager")) {
  install.packages("BiocManager", dependencies = TRUE)
}

if (!require("EnhancedVolcano")) {
  BiocManager::install("EnhancedVolcano")
  library(EnhancedVolcano)
} 

if (!require("gridExtra")) {
  install.packages("gridExtra", dependencies = TRUE)
  library(gridExtra)
}

# Make a list of all the .csv files in your working directory just so you can see what's going on
X = list.files(".",".csv")

# Read the raw data files

results <- read.csv("CKOvsWT_results_phospho.csv", stringsAsFactors=FALSE)

```

```{r}
# First, plot a volcano to visualise the whole dataset

plt <- EnhancedVolcano(results,
    lab = as.vector(results$gene_names),
    x = 'log_fold_change',
    y = 'BH_pVal',
    ylab = bquote(~-Log[10]~italic(q)),
    axisLabSize = 14,
    title = NULL, # Title is optional
    subtitle = NULL,
    caption = NULL,
    xlim = c(min(results$log_fold_change), max(results$log_fold_change)), # Optimise axis limits
    ylim = c(0, max(-log10(results$BH_pVal))),
    pCutoff = 0.05,
    FCcutoff = log2(1), # not using a fold-change cutoff here!
    cutoffLineType = 'blank',
    cutoffLineCol = 'red',
    col = c("grey", "grey", "grey", "red"),
    shape = 16,
    pointSize = 4,
    labSize = 5,
    gridlines.major = FALSE,
    gridlines.minor = FALSE,
    legendPosition = "top",
    legendLabSize = 14,
    legendLabels=c(NA,'q>0.05', NA, 'q<=0.05'),
    legendIconSize= 5) +
    theme(panel.border = element_blank(), #make the annoying colour theme go away
          panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          panel.background = element_blank(),
          axis.ticks = element_line(colour = "black"),
          axis.line = element_line(colour = "black"),
          axis.ticks.length=unit(0.1, "cm"),
          axis.text=element_text(colour="black", size = 12, family = "sans"), 
          legend.background = element_blank(),
          legend.key = element_blank()
          )

plt

ggsave("volcano_total_phospho.pdf", plot = plt, width=8, height = 9)

```
